+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Next Release
 
+### New
+* Allow to UPDATE_TAGS:
+set this ENV var and lobanov will update tags of operations.
+
 ## 0.8.0 (2023-03-08)
 ### New
 * Group related endpoints with tags